+++ packages/codegen/src/generate-descriptors.ts
@@ -319,11 +319,32 @@ export const generateDescriptors = (
     callInterface ? `${prefix}CallData` : null,
   ].filter((v) => v !== null)
 
+  // Going through base64 conversion instead of using binary loader because of esbuild issue
+  // https://github.com/evanw/esbuild/issues/3894
   const imports = `import {${clientImports.join(", ")}} from "${paths.client}";
   import {${commonTypeImports.join(", ")}} from "${paths.types}";
 
+  const toBinary = (() => {
+    const table = new Uint8Array(128);
+    for (let i = 0; i < 64; i++) table[i < 26 ? i + 65 : i < 52 ? i + 71 : i < 62 ? i - 4 : i * 4 - 205] = i;
+    return (base64: string) => {
+      const n = base64.length,
+        bytes = new Uint8Array((n - Number(base64[n - 1] === '=') - Number(base64[n - 2] === '=')) * 3 / 4 | 0);
+      for (let i2 = 0, j = 0; i2 < n;) {
+        const c0 = table[base64.charCodeAt(i2++)], c1 = table[base64.charCodeAt(i2++)];
+        const c2 = table[base64.charCodeAt(i2++)], c3 = table[base64.charCodeAt(i2++)];
+        bytes[j++] = c0 << 2 | c1 >> 4;
+        bytes[j++] = c1 << 4 | c2 >> 2;
+        bytes[j++] = c2 << 6 | c3;
+      }
+      return bytes;
+    };
+  })();
+
   const descriptorValues = import("${paths.descriptorValues}").then(module => module["${prefix}"]);
-  const metadataTypes = import("${paths.metadataTypes}").then(module => 'default' in module ? module.default : module);
+  const metadataTypes = import("${paths.metadataTypes}").then(
+    module => toBinary('default' in module ? module.default : module)
+  );
   `
 
   const descriptorTypes = `${imports}