+++ packages/client/CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Fixed
 
 - codegen: reduced startup memory usage due to esbuild issue [#711](https://github.com/polkadot-api/polkadot-api/pull/711)
+- `polkadot-sdk-compat`: Fix small bug with the unpin-hash enhancer
 
 ## 1.2.1 - 2024-09-10