+++ packages/client/CHANGELOG.md
@@ -15,6 +15,7 @@
 - codegen: reduced startup memory usage due to esbuild issue [#711](https://github.com/polkadot-api/polkadot-api/pull/711)
 - `polkadot-sdk-compat`: Fix small bug with the unpin-hash enhancer
 - `json-rpc-proxy`: Added a property to identify internal "stop" events.
+- `ws-provider`: Some logs when there is a transport close/error
 
 ## 1.2.1 - 2024-09-10